# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Delapre Abbey Events API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "starts_at": { "type": "string", "format": "date-time" },
          "ends_at": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "is_free": { "type": "boolean" },
          "capacity": { "type": "integer" },
          "spots_left": { "type": "integer" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "email": { "type": "string" },
              "first_name": { "type": "string" },
              "last_name": { "type": "string" }
            }
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "status": { "type": "string" },
          "guest_count": { "type": "integer" },
          "guest_names": { "type": "array", "items": { "type": "string" } },
          "booked_at": { "type": "string", "format": "date-time" },
          "cancelled_at": { "type": "string", "format": "date-time", "nullable": true },
          "event": { "$ref": "#/components/schemas/Event" }
        }
      }
    }
  },
  "paths": {
    "/api/events": {
      "get": {
        "summary": "List upcoming events",
        "parameters": [
          {
            "name": "free",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter for free events when set to 1/true"
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Event" }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{eventId}": {
      "get": {
        "summary": "Get event details",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "404": { "description": "Event not found" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "summary": "List user bookings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Booking" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a booking",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_id": { "type": "integer" },
                  "guest_count": { "type": "integer" },
                  "guest_names": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          }
        }
      }
    },
    "/api/bookings/history": {
      "get": {
        "summary": "View booking history",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Booking history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Booking" }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{bookingId}": {
      "delete": {
        "summary": "Cancel a booking",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          }
        }
      }
    }
  }
}